requirements.txt
h5py==3.11.0
huggingface-hub==0.24.6
datasets==2.21.0
hydra-core==1.3.2
lightning==2.5.0.post0
nibabel==4.0.2
nilearn==0.10.4
nsd_access @ git+https://github.com/tknapen/nsd_access@8d9f2ec4fed2516cdda1ad5ebd9fcd7eb99e7b4d
numpy==1.24.4
pandas==2.2.2
peft==0.12.0
pycocoevalcap==1.2
pycocotools==2.0.8
pycortex==1.2.11
safetensors==0.4.4
sentencepiece==0.2.0
scikit-learn==1.5.2
scipy==1.14.1
torch==2.2.2
torchmetrics==1.4.1
textdistance==4.6.3
torchvision==0.17.2
wandb==0.17.7
transformers==4.44.2
random-fourier-features-pytorch==1.0.1